move store definition to dedicated file
